+++ src/audible/__init__.py
@@ -3,7 +3,7 @@ from importlib.metadata import version
 from ._logging import log_helper
 from .auth import Authenticator
 from .client import AsyncClient, Client
-from .login_service import LoginResult, LoginService
+from .login import LoginResult, LoginService
 from .register import RegistrationResult, RegistrationService